utils/scripts/api-generator/pyproject.toml
[tool.poetry]
name = "api-generator"
version = "0.1.0"
description = "Autogenerate markdown files from Crossplane CRDs"
authors = ["Pete Lumbis <pete@upbound.io>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
PyYAML = "^6.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
